Split large views.py into directory. Use wildcard import of views to save code
